backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-google-genai
chromadb
pypdf
python-dotenv
pdf2image
pytesseract
Pillow
